# dist/api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApplicationUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"language": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Parameter": {
				"properties": {
					"property": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"property",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Header": {
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Endpoint": {
				"properties": {
					"host": {
						"type": "string"
					},
					"method": {
						"type": "string"
					},
					"requestLine": {
						"type": "string"
					},
					"queryParameters": {
						"items": {
							"$ref": "#/components/schemas/Parameter"
						},
						"type": "array"
					},
					"headers": {
						"items": {
							"$ref": "#/components/schemas/Header"
						},
						"type": "array"
					},
					"body": {},
					"secured": {
						"type": "boolean"
					},
					"authenticationHeader": {
						"type": "string"
					},
					"authenticationAction": {
						"$ref": "#/components/schemas/Endpoint"
					}
				},
				"required": [
					"host",
					"method",
					"requestLine"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Order": {
				"properties": {
					"command": {
						"type": "string"
					},
					"action": {
						"$ref": "#/components/schemas/Endpoint"
					}
				},
				"required": [
					"command",
					"action"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "custom_voice_commands",
		"version": "1.0.0",
		"description": "<b>This API facilitates custom voice commands to execute the corresponding configured API endpoint</b>",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/register": {
			"post": {
				"operationId": "RegisterAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplicationUser"
							}
						}
					}
				}
			}
		},
		"/admin/commands/create": {
			"post": {
				"operationId": "CreateCommand",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Admin Commands"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Order"
							}
						}
					}
				}
			}
		},
		"/admin/commands/update/{orderId}": {
			"patch": {
				"operationId": "UpdateCommand",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Admin Commands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Order"
							}
						}
					}
				}
			}
		},
		"/admin/commands/delete/{orderId}": {
			"delete": {
				"operationId": "DeleteCommand",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Admin Commands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/commands/execute": {
			"post": {
				"operationId": "ExecuteCommand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "<b>Upload a command as a mono recording formatted to a wav file with 16kHz</b> <br> <b>The endpoint corresponding to the command will be executed.</b>",
				"tags": [
					"Commands"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"voiceCommand": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplicationUser"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}